tests/wpt/web-platform-tests/offscreen-canvas/the-canvas-state/2d.state.saverestore.stack.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.state.saverestore.stack</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.state.saverestore.stack</h1>
<p class="desc">save()/restore() can be nested as a stack</p>


<script>
var t = async_test("save()/restore() can be nested as a stack");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.lineWidth = 1;
ctx.save();
ctx.lineWidth = 2;
ctx.save();
ctx.lineWidth = 3;
_assertSame(ctx.lineWidth, 3, "ctx.lineWidth", "3");
ctx.restore();
_assertSame(ctx.lineWidth, 2, "ctx.lineWidth", "2");
ctx.restore();
_assertSame(ctx.lineWidth, 1, "ctx.lineWidth", "1");

t.done();

});
</script>
